Extract hook to check login state
